+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - 使用 Node.js 與 Angular 5，建立模仿Amazon的網站
 - 安裝 [nodemon](https://github.com/remy/nodemon/) 來啟動 node 程式，參考[介紹](https://oranwind.org/node-nodemon/)
 - [Added] package.json, npm init and added express, body-parser, morgan and mongoose
+- [Added] server/server.js, setup middleware